Refactored user related code
